+++ NEWS.md
@@ -5,6 +5,8 @@ NEXT RELEASE
     character based and works just like grep
   - Two new options from grep: `-A/--after-context` and `-B/--before-context`
   - Lines with multiple matches are now printed only once
+  - Optional caching of pdf-text for faster operation (by Christian Dietrich)
+    This adds a **new dependency**: libgcrypt
   - Bash-completion improvements (by Rainer Müller)
   - Bugfix: Fixed string search (`-F`) now works as advertised with multiple
     patterns